chore: 처치 시 점수 int -> float 변경

+++ Assets/02.Scripts/Enemy/Data/EnemyEvent.cs
@@ -4,7 +4,7 @@ public readonly struct EnemyKilledEvent
 {
     public readonly EnemyBase Enemy;
     public readonly int Score;
-    public readonly int Playtime;
+    public readonly float Playtime;
 
     public EnemyKilledEvent(EnemyBase enemy)
     {
